feat: Update language SVG files for new language support
This commit adds new SVG files for the F#, Elm, Bash, Crystal, Kotlin, Haskell, Zig, Nim, ATS, and Idris programming languages. These SVG files are used for displaying language icons in the editor. The addition of these files improves the visual representation of the supported languages in the code editor.